Allow `pyraise` to include backtrace from Julia
Instead of using `catch_backtrace` in `showerror_string`, which may not always
be called from a catch block with a current exception set, I have added an
optional argument to `pyraise` which can be used to pass backtraces to
`showerror_string`.

Backtraces are only used if exception is not a PyError exception. By default, no
backtrace is shown.

All try-catch blocks in PyCall attempt to pass the backtrace for the current
exception to `pyraise`.