Pull `Py_CompileString` out of `disable_sigint`
Since `Py_CompileString` is unlikely to call Julia callback, there is
no need to put it in `disable_sigint`.  `@pycheckn` for `PyEval_EvalCode`
is in `disable_sigint` because `PyErr_NormalizeException` may call the
constructor of the exception class which may invoke Julia callbacks.

This change may not be necessary but it seems to help Julia compiler
inlining `ccall`.